requirements.txt
Django>=3.1.4
Pillow>=8.0.1
pip>=20.2.4
pytz>=2020.4
sqlparse>=0.4.1
mysqlclient>=2.0.1
asgiref>=3.3.1
pilkit>=2.0
certifi>=2020.12.5
chardet>=3.0.4
urllib3>=1.26.2
six>=1.15.0
django-appconf>=1.0.4
Pygments>=2.7.3
Markdown>=3.0.1
idna>=2.10
setuptools>=50.3.2
pathlib2>=2.3.5